restructure, fix big endian bake
